Merge with updated framework code from Fiji MVC Framework: https://github.com/fijiwebdesign/framework.git https://code.google.com/p/fiji-mvc-framework/
Some bugfixes to Fiji Webmail such as:
Addresses are now parsed correctly when composing.
SMTP transport sending of message now fixed.
Creation of required folders if they don't exist.

Lot's of improvements to the MVC framework.

Added ModelCollection class for Collections of models instead of using
DomainCollection directly.

Added Events to the Model and ModelColletion classes so that extending classes
can subscribe to events such as onSave(), onFind() and onDelete()

Started on the Router which is essential for SEF URLs. Router will allow custom
URL mapping to controllers.