Added keyboard shortcuts. @todo map to keys defined by Gmail or other popular provider for user convenience.
Fixed bug in remove Labels that are IMAP protocol defined Flags

Removed navigation widget that was in app\mail\view\widget\navigation\. This would cause bugs in importing from Git as this was removed locally in dev and not updated to Git.

Adde widget\documentHead which is attached to Factory::getDocument()->head and allows adding of CSS and JS to HTML <head> of current page.