Fixed bug #20 Bootstrap not working
Navigation is preloading the email compose widget modal. When email account is not setup properly we are redirecting to login page which renders and new jquery library into the modal and thus the page. This overwrites the jQuery global making it lose the Bootstrap jquery plugins already set to $.

App->redirect() will now use JS if the Request->siteTemplate is "app" or "ajax" and we are serving HTML.
Login now uses main page template instead of it's own.